Extracting case select reducer test and creating a directory for reducer tests
