+++ client/test/karma/reader/reducers/reducer-test.js
@@ -1,42 +1,10 @@
 /* eslint-disable max-lines */
 import { expect } from 'chai';
-import { reducer, initialState } from '../../../app/reader/reducer';
-import { caseSelectReducer, initialState as caseSelectInitialState } from '../../../app/reader/CaseSelect/CaseSelectReducer';
-import { setViewingDocumentsOrComments } from '../../../app/reader/DocumentList/DocumentListActions';
-import * as Constants from '../../../app/reader/constants';
-import * as CaseSelectConstants from '../../../app/reader/CaseSelect/actionTypes';
-import { caseSelectAppeal } from '../../../app/reader/CaseSelect/CaseSelectActions';
+import { reducer, initialState } from '../../../../app/reader/reducer';
+import { setViewingDocumentsOrComments } from '../../../../app/reader/DocumentList/DocumentListActions';
+import * as Constants from '../../../../app/reader/constants';
 
 /* eslint-disable no-undefined */
-
-describe('CaseSelect reducer', () => {
-  const reduceActions = (actions, state) => actions.reduce(caseSelectReducer, caseSelectReducer(state, {}));
-  
-  it('updates assignment as viewed', () => {
-    const vacolsId = 1;
-    const assignments = [{
-      viewed: false,
-      vacols_id: vacolsId
-    }];
-
-    const state = reduceActions([
-      {
-        type: CaseSelectConstants.RECEIVE_ASSIGNMENTS,
-        payload: {
-          assignments
-        }
-      },
-      {
-        type: CaseSelectConstants.SET_VIEWED_ASSIGNMENT,
-        payload: {
-          vacolsId
-        }
-      }
-    ], caseSelectInitialState);
-
-    expect(state.assignments[0].viewed).to.deep.equal(true);
-  });
-});
 
 describe('Reader reducer', () => {
